Move curtailment methodology to flex_opt
